shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Enterprise shared library for Python microservices - providing exceptions, constants, utilities, observability, resilience, authentication, and database patterns"
authors = [
    {name = "Nam Tran", email = "nam.tranphuong@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["microservices", "enterprise", "shared-library", "fastapi", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

# Core dependencies - minimal by default
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
# Settings management
settings = [
    "pydantic-settings>=2.10.1,<3.0.0",
]

# Observability stack
observability = [
    "structlog>=24.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "prometheus-client>=0.19.0",
]

# Database layer
database = [
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0",
]

# Authentication & security
auth = [
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "argon2-cffi>=23.1.0",
    "cryptography>=41.0.0",
]

# gRPC & Protocol Buffers
grpc = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

# Messaging & Events
messaging = [
    "aiokafka>=0.10.0",
    "aio-pika>=9.3.0",
]

# Caching
caching = [
    "redis>=5.0.0",
    "cachetools>=5.0.0",
]

# HTTP client
http = [
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
]

# FastAPI utilities
fastapi = [
    "fastapi>=0.115.0,<1.0.0",
]

# All optional dependencies
all = [
    "shared[settings,observability,database,auth,grpc,messaging,caching,http,fastapi]",
]

# ============================================================================
# Build System - Modern hatchling with src layout
# ============================================================================

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests/",
    ".venv/",
    "__pycache__/",
    ".pytest_cache/",
    "*.pyc",
]

# ============================================================================
# Dependency Groups (uv native)
# ============================================================================

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.92.0",
    "faker>=22.0.0",
    "factory-boy>=3.3.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    # Include test dependencies for optional features
    "aiosqlite>=0.19.0",
]

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "B007",    # loop control variable not used
    "B008",    # function call in defaults (used for Depends)
    "B017",    # pytest.raises(Exception)
    "B024",    # abstract class without abstract methods
    "B039",    # mutable ContextVar default
    "B904",    # raise from err
    "B905",    # zip without strict=
    "W291",    # trailing whitespace
    "W293",    # blank line with whitespace
    "SIM102",  # nested if
    "SIM103",  # return bool condition
    "SIM105",  # use contextlib.suppress
    "SIM108",  # ternary operator
    "SIM110",  # use any() instead of for loop
    "PLR0913", # too many arguments
    "PLR1714", # merge comparisons
    "RUF005",  # concatenation instead of spread
    "RUF022",  # __all__ not sorted (we group by category)
    "UP046",   # use type parameters (python 3.12+)
    "UP047",   # generic function type parameters
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004", "F841", "PLC0415", "PLR", "B008"]
"src/*" = ["TC002", "TC003", "F841"]  # type-checking imports and context manager variables
"src/**/__init__.py" = ["F401"]  # re-exports in __init__.py files

[tool.ruff.lint.isort]
known-first-party = ["shared"]

[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p no:warnings",
]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require external services)",
    "slow: Slow tests (skip with -m 'not slow')",
]

[tool.coverage.run]
source = ["src/shared"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "@overload",
]
fail_under = 80
show_missing = true
